requirements.txt
langchain
langchain-community
langchain-core
openai
pypdf
tiktoken
aiofiles
fastapi[standard]
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv